% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_all_packages.r
\name{detach_all_packages}
\alias{detach_all_packages}
\title{Detach all packages}
\usage{
detach_all_packages()
}
\description{
This function detaches all packages from the environment.
}
\examples{
detach_all_packages()
}

