% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.plot.r
\name{report.plot}
\alias{report.plot}
\title{Report Plot}
\usage{
report.plot(ggplot_object, file = "")
}
\arguments{
\item{ggplot_object}{ggplot object to be plotted.}

\item{file}{(optional) File name for the file to be saved.  If not supplied
then the data frame's name is used.  Note, ".csv" is appended automatically.}
}
\description{
Creates a plot for an exported report.
Writes a ggplot object to a plot using plot() or writes the plot to a PNG file.
}
\examples{
test <- ggplot(
       keywords_weekly_conversion_metrics, 
       aes(x=week, y=est_search_impression_share, fill=campaign_name)) +
       geom_bar(stat="identity", position="dodge") +
       ggtitle("AdWords - Weekly Impression Share by Geo") + 
       ylim(0, 1) +
       facet_wrap(~keyword, ncol=2)
     )
report.plot(test, "example.png") # Writes to file "example.png"
report.plot(test, "") # Writes to file test.png
report.plot(test, NULL) # Plots without writing to file.
}

