% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_from_file.r
\name{string_from_file}
\alias{string_from_file}
\title{String From File}
\usage{
string_from_file(file_path)
}
\arguments{
\item{file_path}{Path to the file containing the string.}
}
\description{
Reads a string literal from a file and performs string
interpolation on values from the string using `@{}` syntax. 
(e.g @{start_date})
}
\details{
Should consider using Whisker... https://github.com/edwindj/whisker
}
\examples{
string_from_file('./transactions_query.sql')
}

